backend/Authentication/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0